Added support for objects as annotations parameters
