+++ packages/core/input/fetch.ts
@@ -1,8 +1,6 @@
 /**
- * $HttpApi
- * $HttpEvent(method: POST, path: users)
- * $HttpEvent(method: PUT, path: update)
- * $HttpEvent(method: DELETE, path: delete)
+ * $HttpApi(method: "POST", path: "/users/create")
+ * $HttpApi(method: "PUT", path: 3)
  */
 export function createUser(event: any, context: any) {
   return {
@@ -12,8 +10,9 @@ export function createUser(event: any, context: any) {
 }
 
 /**
- * $HttpApi
- * $HttpEvent(method: POST, path: users)
+ * $HttpApi(method: "GET", path: "/users/{id}")
+ * $HttpApi(method: "DELETE", path: (a, b) => a+b)
+ * $HttpApi(method: "DELETE", path: { a: { b: 3 } })
  */
 export function getUser(event: any, context: any) {
   return {
